yes24_daily_briefing/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.2
pandas>=2.1.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.30.0
